Backend/deeplearning_service/requirements.txt
python-dotenv==1.0.1
psycopg2-binary
numpy==1.26.4
kafka-python==2.0.2
pydantic
